+++ tests/testthat/test-equilibrium.R
@@ -3,7 +3,7 @@ context("Equilibrium Model's Tests\n")
 # Estimation setup
 parameters <- list(
   nobs = 1000, tobs = 3,
-  alpha_d = -1.7, beta_d0 = 14.9, beta_d = c(2.3, -1.2), eta_d = c(-1.3, -1.1),
+  alpha_d = -1.9, beta_d0 = 14.9, beta_d = c(2.3, -1.2), eta_d = c(-1.3, -1.1),
   alpha_s = 1.6, beta_s0 = 10.2, beta_s = c(-1.3), eta_s = c(2.5, 2.2),
   sigma_d = 1.0, sigma_s = 1.0, rho_ds = 0.0
 )
@@ -46,12 +46,11 @@ test_that(paste0(
       standard_errors = c("id")
     )
   )
-
   expect_is(est@fit, "list")
 })
 
 test_that(paste0(name(mdl), " fit can be summarized"), {
-  test_summary(est, 41)
+  test_summary(est, 43)
 })
 
 test_that(paste0(
@@ -83,7 +82,7 @@ test_that(paste0("Second stage of '", name(mdl), "' can be estimated"), {
 })
 
 test_that(paste0(name(mdl), " regressions can be summarized"), {
-  test_summary(reg, 88)
+  test_summary(reg, 93)
 })
 
 test_that(paste0(
@@ -94,13 +93,16 @@ test_that(paste0(
 })
 
 test_that(paste0("Optimization of '", name(mdl), "' using GSL succeeds"), {
-  mll <<- maximize_log_likelihood(mdl,
-    start = NULL, step = 1e-2,
-    objective_tolerance = 1e-4,
-    gradient_tolerance = 1e-3,
-    max_it = 1e+3
+  mll <<- estimate(
+    mdl,
+    optimizer = "gsl", control = list(
+      step = 1e-2,
+      objective_tolerance = 1e-4,
+      gradient_tolerance = 1e-3,
+      maxit = 1e+3
+    )
   )
-  testthat::expect_length(mll, 8)
+  expect_is(est@fit, "list")
 })
 
 test_that(paste0(